+++ tests/unit/bases/test_get_base.py
@@ -16,25 +16,58 @@
 #
 
 import pytest
-from craft_providers.bases import get_base_alias, get_base_from_alias, ubuntu
+from craft_platforms import DistroBase
+from craft_providers.bases import (
+    BaseName,
+    almalinux,
+    centos,
+    get_base_alias,
+    get_base_from_alias,
+    ubuntu,
+)
 from craft_providers.errors import BaseConfigurationError
 
 
-def test_get_base_alias():
-    assert get_base_alias(("ubuntu", "22.04")) == ubuntu.BuilddBaseAlias.JAMMY
+@pytest.mark.parametrize(
+    ("base_name", "expected"),
+    [
+        (("ubuntu", "22.04"), ubuntu.BuilddBaseAlias.JAMMY),
+        (("ubuntu", "24.04"), ubuntu.BuilddBaseAlias.NOBLE),
+        (BaseName("ubuntu", "devel"), ubuntu.BuilddBaseAlias.DEVEL),
+        (DistroBase("centos", "7"), centos.CentOSBaseAlias.SEVEN),
+        (DistroBase("centos", "7.4"), centos.CentOSBaseAlias.SEVEN),
+    ],
+)
+def test_get_base_alias(base_name, expected):
+    assert get_base_alias(base_name) == expected
 
 
-def test_get_base_alias_does_not_exist():
-    with pytest.raises(BaseConfigurationError) as exc_info:
-        get_base_alias(("ubuntu", "8.04"))
+@pytest.mark.parametrize(
+    "base_name",
+    [
+        ("ubuntu", "8.04"),
+        ("debian", "1"),
+        ("ubuntu", "24"),
+        ("centos", "6"),
+    ],
+)
+def test_get_base_alias_does_not_exist(base_name):
+    with pytest.raises(
+        BaseConfigurationError, match=r"^Base alias not found for BaseName\(name='"
+    ):
+        get_base_alias(base_name)
 
-    assert exc_info.value == BaseConfigurationError(
-        brief="Base alias not found for BaseName(name='ubuntu', version='8.04')"
-    )
 
-
-def test_get_base_from_alias():
-    assert get_base_from_alias(ubuntu.BuilddBaseAlias.JAMMY) == ubuntu.BuilddBase
+@pytest.mark.parametrize(
+    ("alias", "expected"),
+    [
+        *((alias, ubuntu.BuilddBase) for alias in ubuntu.BuilddBaseAlias),
+        *((alias, centos.CentOSBase) for alias in centos.CentOSBaseAlias),
+        *((alias, almalinux.AlmaLinuxBase) for alias in almalinux.AlmaLinuxBaseAlias),
+    ],
+)
+def test_get_base_from_alias(alias, expected):
+    assert get_base_from_alias(alias) == expected
 
 
 def test_get_base_from_alias_does_not_exist():
@@ -42,5 +75,5 @@ def test_get_base_from_alias_does_not_exist():
         get_base_from_alias("ubuntu 8")  # type: ignore
 
     assert exc_info.value == BaseConfigurationError(
-        brief="Base not found for alias ubuntu 8"
+        brief="Base not found for alias 'ubuntu 8'"
     )