+++ docs/reference/changelog.rst
@@ -4,6 +4,11 @@ Changelog
 See the `Releases page`_ on GitHub for a complete list of commits that are
 included in each version.
 
+X.Y.Z (YYYY-Mon-DD)
+-------------------
+
+- ``bases.BaseName`` is now compatible with the ``craft-platforms.DistroBase``.
+
 2.2.0 (2025-Jan-16)
 -------------------
 - ``hookutil.py`` now available for dependent projects to clean up lxd